# docs/swagger.yaml
basePath: /
definitions:
...
  dto.TranslatedByAuthorCreateRequest:
    properties:
      notes:
        type: string
      original_author_name:
        type: string
      original_language:
        type: string
      work_title:
        type: string
    required:
    - original_author_name
    - original_language
    - work_title
    type: object
  dto.TranslatedByAuthorListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TranslatedByAuthorResponse'
        type: array
      status:
        type: string
      total:
        type: integer
    type: object
  dto.TranslatedByAuthorResponse:
    properties:
      id:
        type: string
      notes:
        type: string
      original_author_name:
        type: string
      original_language:
        type: string
      work_title:
        type: string
    type: object
  dto.TranslatedIntoLanguageCreateRequest:
    properties:
      language_name:
        type: string
      notes:
        type: string
      translator_name:
        type: string
      work_title:
        type: string
    required:
    - language_name
    - translator_name
    - work_title
    type: object
  dto.TranslatedIntoLanguageListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TranslatedIntoLanguageResponse'
        type: array
      status:
        type: string
      total:
        type: integer
    type: object
  dto.TranslatedIntoLanguageResponse:
    properties:
      id:
        type: string
      language_name:
        type: string
      notes:
        type: string
      translator_name:
        type: string
      work_title:
        type: string
    type: object
...
  handler.JSONResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for Hemra Şirow literary and biography website
  title: Literary Backend API
  version: "1.0"
paths:
...
  /api/translations/by-author:
    get:
      description: List all translations by author
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TranslatedByAuthorListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: List translations by author
    post:
      consumes:
      - application/json
      description: Create a new translation by author
      parameters:
      - description: Translation by author data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TranslatedByAuthorCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TranslatedByAuthorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Create translation by author
  /api/translations/by-author/{id}:
    delete:
      description: Delete an existing translation by author by id
      parameters:
      - description: Translation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Delete translation by author
    get:
      description: Get a single translation by author by id
      parameters:
      - description: Translation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TranslatedByAuthorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Get translation by author
  /api/translations/into-language:
    get:
      description: List all translations into language
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TranslatedIntoLanguageListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: List translations into language
    post:
      consumes:
      - application/json
      description: Create a new translation into language
      parameters:
      - description: Translation into language data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TranslatedIntoLanguageCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TranslatedIntoLanguageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Create translation into language
  /api/translations/into-language/{id}:
    delete:
      description: Delete an existing translation into language by id
      parameters:
      - description: Translation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Delete translation into language
    get:
      description: Get a single translation into language by id
      parameters:
      - description: Translation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TranslatedIntoLanguageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Get translation into language
...
schemes:
- http
- https
swagger: "2.0"